services/merlin-api/requirements.txt
fastapi
uvicorn
pandas
python-multipart

fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
pydantic-settings>=2
numpy
pandas
scikit-learn
faiss-cpu
psycopg[binary]
SQLAlchemy

implicit
scipy
faiss-cpu
implicit==0.7.2